requirements.txt
langchain>=0.3.20
langchain-community>=0.3.19
langchain-text-splitters>=0.3.6
langchain-chroma>=0.2.2
langchain-core>=0.3.43
langchain-google-genai>=0.0.10
pandas>=2.2.3
python-dotenv>=1.0.1
fastapi>=0.110.0
uvicorn>=0.29.0
sqlalchemy>=2.0.29
requests>=2.32.3
chromadb>=0.5.0
pydantic>=2.10.6
tqdm>=4.66.0
google-generativeai>=0.3.2
flask>=2.0.0
flask-cors>=3.0.10
langchain-openai>=0.0.5
steam>=1.4.4